refactor: organise class into respective packages
'task', 'storage', 'command'
