+++ src/main/java/duke/command/CommandFactory.java
@@ -1,3 +1,5 @@
+package duke.command;
+
 import java.util.Map;
 import java.util.HashMap;
