+++ src/main/java/duke/command/UnknownCommandException.java
@@ -1,3 +1,5 @@
+package duke.command;
+
 public class UnknownCommandException extends Exception {
     public UnknownCommandException(String type) {
         super("Unknown command provided '" + type + "'");
