+++ src/main/java/duke/storage/CorruptSaveException.java
@@ -1,3 +1,5 @@
+package duke.storage;
+
 public class CorruptSaveException extends Exception {
     public CorruptSaveException(String line) {
         super("Save file is corrupted at line '" + line + "'");
