+++ src/main/java/duke/task/EventParser.java
@@ -1,7 +1,11 @@
+package duke.task;
+
 import java.util.regex.Pattern;
 import java.util.regex.Matcher;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
+
+import duke.storage.CorruptSaveException;
 
 public class EventParser extends TaskParser {
     @Override
