+++ src/main/java/duke/task/TaskParser.java
@@ -1,3 +1,7 @@
+package duke.task;
+
+import duke.storage.CorruptSaveException;
+
 public abstract class TaskParser {
     public abstract Task parse(String description) throws IncompleteTaskException;
     public abstract Task parseFromFile(String description) throws CorruptSaveException;
