+++ src/main/java/duke/task/TaskParserFactory.java
@@ -1,3 +1,5 @@
+package duke.task;
+
 import java.util.Map;
 import java.util.HashMap;
