+++ src/main/java/duke/task/ToDo.java
@@ -1,3 +1,4 @@
+package duke.task;
 
 public class ToDo extends Task {
     public ToDo(String description) {
