+++ src/main/java/duke/task/UnknownTaskTypeException.java
@@ -1,3 +1,5 @@
+package duke.task;
+
 public class UnknownTaskTypeException extends Exception {
     public UnknownTaskTypeException(String type) {
         super("Unknown task type specified '" + type + "'");
